Removed id from GObj. Replaced all pointers to smart pointers. Updated tests.

+++ src/gltools.cpp
@@ -19,7 +19,7 @@ std::vector<Vertex> circleTriangleFan(R r, int count) {
     if (count < 3)
         throw too_few_vertices();
     std::vector<Vertex> vertexList(count + 2);
-    double step = M_PI * 2 / count;
+    float step = M_PI * 2 / count;
     vertexList[0] = Vertex({0,0});
     for (int i = 0; i < count + 1; i++)
         vertexList[i + 1] = {std::cos(step * i) * r, std::sin(step * i) * r};
